Refactor project structure: remove unused files, add new configuration and main execution files, and update dependencies. WE THE BEST MUSIC!!
