+++ .gitignore
@@ -1,5 +1,8 @@
 .env
 __pycache__
+coderefineai_executor/build
+coderefineai_executor/dist
+coderefineai_executor/coderefineai_executor.egg-info
 poetry.lock
 crystalblue
 dataset
