+++ coderefineai_executor/coderefineai_executor/utils.py
@@ -1,3 +1,28 @@
+import base64
+import ast
+from html.parser import HTMLParser
+import re
+
+def extract_class(source_code: str, class_name: str) -> str:
+    tree = ast.parse(source_code)
+    for node in tree.body:
+        if isinstance(node, ast.ClassDef) and node.name == class_name:
+            return ast.unparse(node)  
+
+    return None  
+
+def extract_python_code(text):
+    code_blocks = re.findall(r"```py(.*?)```", text, re.DOTALL)
+    return "\n\n".join(code_blocks).strip()
+def encode_base64(text: str) -> str:
+    encoded_bytes = base64.b64encode(text.encode("utf-8"))
+    return encoded_bytes.decode("utf-8")
+
+def decode_base64(encoded_text: str) -> str:
+    decoded_bytes = base64.b64decode(encoded_text.encode("utf-8"))
+    return decoded_bytes.decode("utf-8")
+
+
 template = """
 import json
 import itertools
@@ -21,7 +46,7 @@ from heapq import heappush, heappop, heapify
 from collections import deque, defaultdict, OrderedDict, Counter
 from typing import List, Optional, Tuple
 from functools import lru_cache
-from sortedcontainers import SortedList
+# from sortedcontainers import SortedList
 
 {import_code}
