--- core/LICENSE
@@ -1,9 +1,0 @@
-
-### Step 6: Create a `LICENSE` File
-
-Choose a license for your package and create a `LICENSE` file. For example, you can use the MIT License.
-
-```text
-MIT License
-
-...
