--- core/executor/__init__.py
@@ -1,5 +1,0 @@
-# core/executor/__init__.py
-
-from .executor import Executor, ExecutorResponse
-
-__all__ = ["Executor", "ExecutorResponse"]
