--- core/executor/utils/base64.py
@@ -1,9 +1,0 @@
-import base64
-
-def encode_base64(text: str) -> str:
-    encoded_bytes = base64.b64encode(text.encode("utf-8"))
-    return encoded_bytes.decode("utf-8")
-
-def decode_base64(encoded_text: str) -> str:
-    decoded_bytes = base64.b64decode(encoded_text.encode("utf-8"))
-    return decoded_bytes.decode("utf-8")
