+++ pyproject.toml
@@ -21,6 +21,7 @@ numpy = "^2.2.2"
 radon = "^6.0.1"
 llm-sandbox = "^0.2.1"
 pandas = "^2.2.3"
+twine = "^6.1.0"
 
 
 [tool.poetry.dependencies.smolagents]
